requirements.txt
requests==2.31.0
openpyxl==3.1.2
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
